climate-mq/requirements.txt
Django==5.0.3
psycopg[binary]==3.1.18
django-filter==23.5
djangorestframework-gis==1.0.0
djangorestframework==3.15.1
pika
djangorestframework-datatables
numpy
torch
huggingface_hub
celery
scikit-learn
pandas
